Move pfss code to separate file
